pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-db-mcp"
version = "1.0.0"
description = "Unified DB MCP Server - schema extraction and migration across multiple databases"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "mcp>=0.9.0",
    "starlette>=0.36.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "mysql-connector-python>=8.2.0",
    "supabase>=2.0.0",
    "pymongo>=4.6.0",
    "pyodbc>=4.0.39",
    "cassandra-driver>=3.29.0",
]

[tool.setuptools]
packages = [
    "unified_db_mcp",
    "unified_db_mcp.tools",
    "unified_db_mcp.database_connectors",
    "unified_db_mcp.helpers",
]

